requirements.txt
# Core LangChain and Hugging Face libraries
langchain
langchain-community
langchain-huggingface
python-dotenv

# Agent and Tools
tavily-python # For the web search tool

# For the LLM backend
torch

# For text-to-speech and language detection
gTTS
langdetect

# For microphone input and speech-to-text
SpeechRecognition
PyAudio # A dependency for SpeechRecognition to access the microphone

# Pinned for Python 3.9 compatibility
aiohttp<3.9.0